[#934] Rename a few files for consistency
* Adds `.fixture` suffix to test fixture files
* Renames files to follow kebab-case for utils
