+++ tests/unit/vue/fixtures/domainBranding.fixture.ts
