+++ tests/unit/vue/fixtures/window.fixture.ts
@@ -1,0 +1,94 @@
+// tests/unit/vue/fixtures/window.fixture.ts
+import { OnetimeWindow } from '@/types/declarations/window';
+
+export const windowFixture: OnetimeWindow = {
+  authenticated: false,
+  baseuri: 'https://dev.onetimesecret.com',
+  cust: null,
+  custid: '',
+  domains_enabled: false,
+  email: 'test@example.com',
+  frontend_host: 'https://dev.onetimesecret.com',
+  locale: 'en',
+  is_default_locale: false,
+  supported_locales: ['en', 'es', 'fr'],
+  ot_version: '0.19.0 (a5ccaf82)',
+  plans_enabled: true,
+  regions_enabled: true,
+  ruby_version: 'ruby-335',
+  shrimp: 'test-shrimp-token',
+  site_host: 'dev.onetimesecret.com',
+  stripe_customer: undefined,
+  stripe_subscriptions: undefined,
+  form_fields: {},
+  authentication: {
+    enabled: true,
+    signup: true,
+    signin: true,
+    autoverify: false,
+  },
+  secret_options: {
+    default_ttl: 604800.0,
+    ttl_options: [60, 3600, 86400, 604800, 1209600, 2592000],
+  },
+  available_plans: {},
+  support_host: 'https://docs.onetime.co',
+  display_links: true,
+  display_masthead: true,
+  metadata_record_count: 0,
+  plan: {
+    identifier: 'anonymous',
+    planid: 'anonymous',
+    price: 0,
+    discount: 0,
+    options: {
+      ttl: 604800.0,
+      size: 100000,
+      api: false,
+      name: 'Anonymous',
+    },
+  },
+  is_paid: false,
+  default_planid: 'basic',
+  received: [],
+  notreceived: [],
+  has_items: false,
+  regions: {
+    enabled: true,
+    current_jurisdiction: 'EU',
+    jurisdictions: [
+      {
+        identifier: 'EU',
+        display_name: 'European Union',
+        domain: 'eu.onetimesecret.com',
+        icon: 'fa6-solid:earth-europe',
+      },
+      {
+        identifier: 'US',
+        display_name: 'United States',
+        domain: 'us.onetimesecret.com',
+        icon: 'fa6-solid:earth-americas',
+      },
+    ],
+  },
+  incoming_recipient: 'incoming@solutious.com',
+  available_jurisdictions: ['EU', 'US'],
+  enjoyTheVue: true,
+  canonical_domain: 'dev.onetimesecret.com',
+  domain_strategy: 'canonical',
+  domain_id: '',
+  display_domain: 'dev.onetimesecret.com',
+  domain_branding: {
+    instructions_pre_reveal: '',
+    font_family: 'sans',
+    instructions_post_reveal: '',
+    image_encoded: '',
+    image_content_type: 'image/png',
+    allow_public_homepage: 'false',
+    instructions_reveal: '',
+    corner_style: 'rounded',
+    button_text_light: 'true',
+    primary_color: '#36454F',
+    image_filename: '',
+  },
+};
