+++ tsconfig.json
@@ -63,7 +63,7 @@
       "@tests/*": ["./tests/*"]
     },
     "typeRoots": ["./node_modules/@types", "./src/types", "./tests/unit/vue"],
-    "types": ["vite/client", "jsdom", "node"]
+    "types": ["jsdom", "node"]
   },
   "include": [
     "**/*.config.*s",
